enhancement: add warning for vote input in voting widget
