+++ packages/react-app-revamp/components/PriceCurve/wrapper.tsx
@@ -1,22 +1,15 @@
 import useContestConfigStore from "@hooks/useContestConfig/store";
-import { useContestVoteTimings } from "@hooks/useContestVoteTimings";
 import { useCurrencyStore } from "@hooks/useCurrency/store";
 import { convertToDisplayPrice } from "@hooks/useCurrency/useDisplayPrice";
 import useNativeRates from "@hooks/useCurrency/useNativeRates";
 import useCurrentPricePercentageIncrease from "@hooks/useCurrentPricePercentageIncrease";
-import { PriceCurveType } from "@hooks/useDeployContest/types";
-import usePriceCurveChartData from "@hooks/usePriceCurveChartData";
-import usePriceCurveMultiple from "@hooks/usePriceCurveMultiple";
-import usePriceCurvePoints from "@hooks/usePriceCurvePoints";
-import usePriceCurveType from "@hooks/usePriceCurveType";
-import usePriceCurveUpdateInterval from "@hooks/usePriceCurveUpdateInterval";
+import usePriceCurveData from "@hooks/usePriceCurveData";
 import { useCountdownTimer } from "@hooks/useTimer";
 import { useParentSize } from "@visx/responsive";
 import { useCallback, useEffect, useMemo } from "react";
-import { useReadContract } from "wagmi";
+import { useShallow } from "zustand/shallow";
 import PriceCurve from "./index";
 import usePriceCurveChartStore from "./store";
-import { useShallow } from "zustand/shallow";
 
 const DEFAULT_CHART_HEIGHT = 300;
 
@@ -39,95 +32,25 @@ const PriceCurveWrapper = ({
 }: PriceCurveWrapperProps) => {
   const { parentRef, width } = useParentSize({ debounceTime: 150 });
   const contestConfig = useContestConfigStore(useShallow(state => state.contestConfig));
-
-  const {
-    voteTimings,
-    isLoading: isTimingsLoading,
-    isError: isTimingsError,
-  } = useContestVoteTimings({
-    address: contestConfig.address,
-    chainId: contestConfig.chainId,
-    abi: contestConfig.abi,
-  });
-
-  const {
-    data: costToVoteRaw,
-    isLoading: isCostLoading,
-    isError: isCostError,
-  } = useReadContract({
-    address: contestConfig.address,
-    chainId: contestConfig.chainId,
-    abi: contestConfig.abi,
-    functionName: "costToVote",
-    query: { staleTime: Infinity },
-  });
-
-  const startPrice = Number(costToVoteRaw ?? 0);
-  const voteStartSec = voteTimings ? Number(voteTimings.voteStart) : 0;
-  const contestDeadlineSec = voteTimings ? Number(voteTimings.contestDeadline) : 0;
-  const startTimeMs = voteStartSec * 1000 + 1000;
-  const endTimeMs = contestDeadlineSec * 1000 + 1000;
-  const totalVotingMinutes = voteTimings ? Math.floor((contestDeadlineSec - voteStartSec) / 60) : 0;
-
   const displayCurrency = useCurrencyStore(state => state.displayCurrency);
   const { data: nativeRates } = useNativeRates();
-  const endTime = useMemo(() => new Date(endTimeMs), [endTimeMs]);
-  const votingTimeLeft = useCountdownTimer(endTime);
 
   const {
-    priceCurveMultiple,
-    isLoading: isMultipleLoading,
-    isError: isMultipleError,
-  } = usePriceCurveMultiple({
-    address: contestConfig.address,
-    abi: contestConfig.abi,
-    chainId: contestConfig.chainId,
-  });
-
-  const {
-    priceCurveUpdateInterval,
-    isLoading: isIntervalLoading,
-    isError: isIntervalError,
-  } = usePriceCurveUpdateInterval({
-    address: contestConfig.address,
-    abi: contestConfig.abi,
-    chainId: contestConfig.chainId,
-  });
-
-  const { priceCurveType } = usePriceCurveType({
-    address: contestConfig.address,
-    abi: contestConfig.abi,
-    chainId: contestConfig.chainId,
-    version: contestConfig.version,
-  });
-
-  const {
-    pricePoints,
-    isLoading: isPointsLoading,
-    isError: isPointsError,
-  } = usePriceCurvePoints({
+    chartData,
+    currentPrice,
+    currentIndex,
     startPrice,
-    multiple: Number(priceCurveMultiple),
+    priceCurveType,
+    priceCurveUpdateInterval,
     startTimeMs,
     endTimeMs,
-    updateIntervalSeconds: priceCurveUpdateInterval,
-    priceCurveType,
-    enabled:
-      !isTimingsLoading &&
-      !isTimingsError &&
-      !isCostLoading &&
-      !isCostError &&
-      !!voteTimings &&
-      startPrice > 0 &&
-      !isMultipleLoading &&
-      !isMultipleError &&
-      !!priceCurveMultiple &&
-      !!priceCurveUpdateInterval &&
-      !isIntervalLoading &&
-      !isIntervalError,
-  });
+    totalVotingMinutes,
+    isLoading,
+    isError,
+  } = usePriceCurveData();
 
-  const { chartData, currentPrice, currentIndex } = usePriceCurveChartData({ pricePoints });
+  const endTime = useMemo(() => new Date(endTimeMs), [endTimeMs]);
+  const votingTimeLeft = useCountdownTimer(endTime);
 
   const { currentPricePercentageData } = useCurrentPricePercentageIncrease({
     address: contestConfig.address,
@@ -167,9 +90,6 @@ const PriceCurveWrapper = ({
     const shouldWarn = showPriceWarning && secondsUntilNextUpdate < 15 && votingTimeLeft > 60;
     setShowPriceUpdateWarning(shouldWarn);
   }, [showPriceWarning, secondsUntilNextUpdate, votingTimeLeft, setShowPriceUpdateWarning]);
-
-  const isLoading = isTimingsLoading || isCostLoading || isMultipleLoading || isIntervalLoading || isPointsLoading;
-  const isError = isTimingsError || isCostError || isMultipleError || isIntervalError || isPointsError;
 
   if (isLoading) {
     return <div ref={parentRef} style={{ height }} className="w-full animate-pulse rounded-lg bg-neutral-2" />;
